Fix #40

+++ app/customer/services.py
@@ -56,7 +56,7 @@ class CustomerService(remote.Service):
         CustomerSearchMessage)
     ID_RESOURCE = endpoints.ResourceContainer(
         message_types.VoidMessage,
-        id=messages.IntegerField(1, variant=messages.Variant.INT32))
+        id=messages.IntegerField(1, variant=messages.Variant.INT64))
 
     @endpoints.method(ID_RESOURCE,
                       CustomerGetMessage,
