+++ app/purchase/services.py
@@ -147,7 +147,7 @@ class PurchaseService(remote.Service):
 
     ID_RESOURCE = endpoints.ResourceContainer(
         message_types.VoidMessage,
-        id=messages.IntegerField(1, variant=messages.Variant.INT32))
+        id=messages.IntegerField(1, variant=messages.Variant.INT64))
 
     @endpoints.method(ID_RESOURCE,
                       message_types.VoidMessage,
