+++ app/supplier/services.py
@@ -136,7 +136,7 @@ class SupplierService(remote.Service):
 
     ID_RESOURCE = endpoints.ResourceContainer(
         message_types.VoidMessage,
-        id=messages.IntegerField(1, variant=messages.Variant.INT32))
+        id=messages.IntegerField(1, variant=messages.Variant.INT64))
 
     @endpoints.method(ID_RESOURCE,
                       message_types.VoidMessage,
